requirements.txt
django
djangorestframework
psycopg2-binary
django_filter
python-dotenv
celery
redis
requests
pytz
pytest-django
model_bakery
pyyaml
